requirements.txt
bcrypt==3.1.6
cffi==1.12.1
Django==2.1.7
ecdsa==0.13
future==0.17.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pycparser==2.19
python-jose==3.0.1
pytz==2018.9
rsa==4.0
six==1.12.0
